refactor: extract PDF source inline evidence
